accounts: move relay stuff to own file
