+++ crates/notedeck/src/account/mod.rs
@@ -1,1 +1,2 @@
 pub mod accounts;
+pub mod relay;
